Launches via CMAKE w/ library and GUI exe as separate targets. move to better package structure for cmake library builds.  gui "app" moved to right directory Still figuring out CMAKE so things like the working directory for VSCode and the ability to copy shader files to the exe output directory are still a bit of a hack... oh well

+++ src/CPU/CPU2A03.cpp
@@ -1,5 +1,5 @@
-#include "cpu2A03.h"
-#include "AddressingModeHandler.h"
+#include <ControlDeck/CPU/cpu2A03.h>
+#include <ControlDeck/CPU/AddressingModeHandler.h>
 #include <string.h>
 namespace NES {
     Cpu2a03::Cpu2a03(Ppu2C02 *ppu, Cartridge *cartridge) : ppu(ppu), cartridge(cartridge) {
@@ -379,7 +379,7 @@ namespace NES {
     *        1. Get data from PC+1
     */
     void Cpu2a03::getImmediateAddress(OpCodeArgs &args) {
-        args.args[0] = (registers.programCounter);
+        args.args[0] = (uint8_t) (registers.programCounter);
     }
 
     /**
